+++ Cmdline/Action/Compat.cs
@@ -295,7 +295,7 @@ namespace CKAN.CmdLine
         {
             foreach (var h in CompatSubOptions.GetHelp(verb))
             {
-                user?.RaiseError(h);
+                user?.RaiseError("{0}", h);
             }
         }