+++ Cmdline/Action/Filter.cs
@@ -224,7 +224,7 @@ namespace CKAN.CmdLine
         {
             foreach (var h in FilterSubOptions.GetHelp(verb))
             {
-                user?.RaiseError(h);
+                user?.RaiseError("{0}", h);
             }
         }