+++ Cmdline/Action/Mark.cs
@@ -147,7 +147,7 @@ namespace CKAN.CmdLine
         {
             foreach (var h in MarkSubOptions.GetHelp(verb))
             {
-                user.RaiseError(h);
+                user.RaiseError("{0}", h);
             }
         }