+++ Cmdline/Action/Search.cs
@@ -27,7 +27,7 @@ namespace CKAN.CmdLine
                 user.RaiseError(Properties.Resources.ArgumentMissing);
                 foreach (var h in Actions.GetHelp("search"))
                 {
-                    user.RaiseError(h);
+                    user.RaiseError("{0}", h);
                 }
                 return Exit.BADOPT;
             }
@@ -117,7 +117,7 @@ namespace CKAN.CmdLine
 
                 foreach (CkanModule mod in matching)
                 {
-                    user.RaiseMessage(mod.identifier);
+                    user.RaiseMessage("{0}", mod.identifier);
                 }
             }